Updating default value for display_errors config option
Previously this default to "On", which it shouldn't have done (as we should design roles to be suitable for production environments by default).

This also caused conflicts with other PHP roles (i.e. apache-php) where this variable was set to the 'correct' option. This lead to multiple change events ocurring whenever the ansible roles were applied.

This change causes both roles to agree and thus resolves this issue.