Adding changes to default PHP ini options to meet security recommendations
Most of these have no impact on us since we aren't using anything that can't use the tightened options, they just aren't set by default.

It is worth highlighting the fopen function is disabled for accessing URLs (again a security recommendation) which might affect people in the future. This can be reversed in a project by overriding the option, but ideally people should use a proper library to get such resources where possible.